fix/folder-stuctures
change folder stucture in create commponents folder to auth.controllers.js,auth.customValidation.js ,auth.route.js , auth.Sevices.js, auth.validation.js add file
