--- auth/login/loginServices.js
@@ -1,61 +1,0 @@
-import userMdl from "../../schema/userMdl.js";
-import jwt from "jsonwebtoken";
-import config from "../../config.js";
-const {
-  accessTokenKey,
-  refreshTokenKey,
-  accessTokenExpiry,
-  refreshTokenExpiry,
-} = config.jwt;
-
-const loginUser = async (reqBody) => {
-  const { email, password } = reqBody;
-  try {
-    // Find user by username
-    const findUser = await userMdl.findOne({ email });
-
-    if (!findUser) {
-      const error = new Error("Invalid credentials");
-      error.status = 400;
-      throw error;
-    }
-
-    if (password !== findUser.password) {
-      const error = new Error("Invalid password");
-      error.status = 400;
-      throw error;
-    }
-    const accessToken = jwt.sign({ _id: findUser._id }, accessTokenKey, {
-      expiresIn: accessTokenExpiry,
-    });
-
-    const refreshToken = jwt.sign({ _id: findUser._id }, refreshTokenKey, {
-      expiresIn: refreshTokenExpiry,
-    });
-    findUser.refreshToken = refreshToken;
-    await findUser.save();
-
-    return {
-      accessToken: accessToken,
-      refreshToken: refreshToken,
-    };
-  } catch (err) {
-    const error = new Error(err.message);
-    throw error;
-  }
-};
-const logoutUser = async (req, res) => {
-  req.session.destroy((err) => {
-    if (err) {
-      const error = new Error("error in logout");
-      error.status = 400;
-      throw error;
-    }
-  });
-  res.clearCookie("connect.sid");
-  res.status(200).send({ message: "logout" });
-};
-export default {
-  loginUser,
-  logoutUser,
-};
