+++ privatePage/pageRoute.js
@@ -3,10 +3,14 @@ import user from "../schema/userMdl.js";
 import express from "express";
 const router = express.Router();
 
-router.get("/", middleware.verify, async (req, res) => {
+router.get("/private", middleware.verify, async (req, res) => {
   const foundUser = await user.findOne({ _id: req.user });
-
-  res.status(200).json({ message: `hello ${foundUser.email} login` });
+  const userProfile = {
+    name: foundUser.name,
+    email: foundUser.email,
+    registeredAt: foundUser.registeredAt,
+  };
+  res.status(200).json(userProfile);
 });
 
 export default router;
